requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
pydantic>=2.8.0
httpx>=0.27.0
python-dotenv>=1.0.0
pytest>=8.0.0